Store data on File when read
This way, an app that uses Bowser doesn't have to keep a reference to
the data. It can simply keep a reference to the file, which may already
be in the `FileList`.

Once the file is GCed, though, if the app doesn't store its own
reference to the data, it will also be GCed.