Add DelegateNative module (#19)
* Add DelegateNative module

This commit extracts the method_missing functionality from Element to
its own module.

This module can be included to delegate unhandled incoming messages to
the native object being wrapped. These messages will be transformed to
match JavaScript conventions and there may be several properties checked
for a single message.

Spec coverage still lives in element_spec.rb. We may want to extract
that, as well, but it's okay to remain there for now.

* Add include_all parameter in `respond_to_missing?`

The include_all flag determines whether to include protected and private
methods, but there is no such thing in JavaScript, so we just ignore
that argument, though we accept it for the purpose of arity.