Support using Promise in place of a JS Promise
According to the spec, promises must be `then`-able in JS terms, so in
Ruby terms that means the `then` ivar needs to be a proc. We also do the
same with `catch` just for good measure.

This commit also adds support for receiving a callable handler as an
argument to `then` or `catch`, allowing it to be a drop-in replacement
for a JS Promise.